backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pyyaml==6.0.1
beautifulsoup4==4.12.3
feedparser==6.0.11
pydantic==2.9.0
